requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
folium==0.20.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kafka-python==2.2.15
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
narwhals==2.1.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.1
streamlit-folium==0.25.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.4.0
